Add met forcing files for updated experiments
@gabsun has set up two experiments on modelevaluation.org to accept
model output from CABLE: a five site test and a forty two site test.

The user now specifies the `experiment` key in the config file which
selects either to run with the five sites, the forty two sites, or a
single site that is a subset of the five sites.

Remove `met_subset` key from config file as met files are specified by
the `experiment` key.

For specifying met sites in each experiment, store only the site id
internally for each met site and glob the `MET_DIR` directory to get
the names of the corresponding met file. This way, when the met
file data sets are updated, only the `MET_DIR` path needs to be updated.
- check site ids map to valid files in `validate_environment()`

Update unit tests for `check_config()`.