Use `TMPDIR` environment variable instead of CWD
Previously, pytest fails if the command is not run from the
project root directory. This change uses the `TMPDIR`
environment variable instead of `Path.cwd()` so that pytest
runs correctly from any directory.

Using `TMPDIR` has the added benefit that we won't run into
permission issues if we run the tests on a central installation of
benchcab.

Fixes #26