Add support for patching derived types
Currently the `patch` method in the f90nml package does not
preserve existing derived type parameters when a derived type
patch is applied.

This change allows for existing derived type parameters to be
preserved when adjusting the namelist file for each task.

Update unit tests for `adjust_namelist_file()` to test existing
namelist parameters are preserved.