Add branch specific namelist settings
This change allows for users to add branch specific
namelist settings in the config file that get "patched"
to the base namelist settings used for both branches.
These additional settings are set by specifying the
`patch` key in each realisation. Currently, this change
supports the ability to add a `patch` to both
realisations.

Addresses issue #29