Add version flag
Currently the version is hard coded in both `setup.cfg` and `meta.yaml`.
Ideally, we want the version number to be defined in a single place in
the code base. This change defines the `setup.cfg` file to be the single
source of truth for the version number.

Use `importlib` to obtain the version number from `setup.cfg` and store in
the `__version__` dunder variable in `benchcab/__init__.py`.

Add a `--version` command line flag to `benchcab` which prints the version
number to stdout.

Use Jinja2 templating with `load_setup_py_data()` to render the version
number in `meta.yaml`. See [conda docs] for details.

Fixes #38

[conda docs]: https://docs.conda.io/projects/conda-build/en/latest/resources/define-metadata.html#templating-with-jinja