Redesign command line interface
Currently, benchcab uses two main program entry points: benchcab and
benchsiterun. Having benchsiterun accessible at the same level as
benchcab allows the user to run benchsiterun incorrectly, for example if
the user tries to run benchsiterun before the work directory is
generated. Ideally benchsiterun is invoked only by the job script
submitted by benchcab and should be "hidden" to the user.

Move the benchsiterun main program entry point into benchcab.

Add the ability to run specific steps in the workflow in isolation via
[subcommands].

These steps include checkout, build, setting up the fluxnet task work
directory and executing the fluxnet tasks.

Add the --no-submit flag to disable the job script submission when
executing fluxnet tasks.

This flag can also be specified in the main workflow if the user wishes
to execute tasks on the login node (if the computation is inexpensive
enough so that the job doesn't get killed).

Add unit tests to test the command line arguments are parsed correctly.

This should provide a command line interface that makes it difficult for
the user to incorrectly run benchsiterun (now `benchcab fluxnet-run-tasks
--no-submit`).

Fixes #21

[subcommands]: https://docs.python.org/3/library/argparse.html#sub-commands