Polish standard output messages
Currently, `benchcab` outputs stdout from other commands such as `svn`
and `./build3.sh`, but does not output anything when running `./cable`
or when generating particular directory structures.

This commit improves the messages output to stdout by `benchcab` for all
sub-commands. We also add a boolean flag to specify verbose output for
each sub-command.

Write the standard output from CABLE to the `tasks/<task_name>/out.txt`
directory instead of writing to `/dev/null` so that it is accessible to
the user.

Fixes #67