Add ability to use branches of other users
The user should be able to checkout branches of other users so that test
suites can be reproduced by any user.

This change lets the user specify the path relative to the SVN root of
the CABLE repository for each branch by specifying the `path` key in the
configuration file. Remove the `trunk` and `share_branch` keys from the
configuration file as this information is contained within `path`.

This change repurposes the `name` key for each branch to be an alias
used internally by benchcab. This lets the user specify the directory
name when checking out the branch, that is, `name` is the optional
`PATH` argument to svn checkout`. This is useful in the case that we
have a collision in branch names when checking out multiple branches.

Remove the `-M` PBS argument for sending emails in the job script as the
default behaviour is sufficient.

Remove the `user` key as this is now redundant.

Update unit tests for `check_config()`.

Fixes #61