Refactor code and unit tests
This change refactors the code to be more object oriented so that we can
better support mocking via dependency injection rather than resorting to
the `unittest.mock.patch` function. This allows us to write unit tests
that are simpler and that preserve the API layer (as opposed to using
`unittest.mock.patch` which breaks the API layer).

Fixes #102