Remove synonymous names for 'fluxsite'
Currently, there is inconsistent naming conventions being used to
describe fluxsite tests which can be confusing for developers and users.
These include names such as 'site', 'single site', 'fluxsite',
'fluxnet', etc. This change replaces all occurrences of these names in
the code base and documentation with 'fluxsite'.

We choose 'fluxsite' as this describes the nature of the tests (as
opposed to 'fluxnet' which describes a product/network).

Fixes #62