Promote PBS constants to optional parameters
This change promotes PBS related constants to optional parameters in the
configuration file so that PBS flags can be set at runtime.

This is useful in running multiple benchcab instances with different job
parameters such as memory and the number of CPUs. This will also allow
us to easily find an optimal number of CPUs to use to maximise
performance.

This change also adds the ability to switch on and off multiprocessing
at runtime via an optional parameter in the config file.

Fixes #104