Run executable from task dir
Currently, the CABLE executable is run from the current working
directory instead of the task directory by using the absolute path to
the executable. However, we should be running the executable from the
task directory (otherwise CABLE will complain it cannot find the soil
and PFT namelist files). This was a bug that was introduced in the #99.

This change makes it so we change into the task directory before running
the CABLE executable.

Fixes #123