Add patch remove feature
Currently, `benchcab` does not support removing namelist parameters from
the base namelist file. For example, legacy CABLE branches may not be
compatible with a namelist parameter in the base namelist file.

This change adds a `patch_remove` key which will remove a given set of
namelist parameters for a specific branch.

Fixes #92