Move experiment key under `fluxsite`
The `experiment` key in the config file is specific to fluxsite tests.
This key is better placed under the `fluxsite` key as we will soon have
the ability to run spatial only tests.

This change moves the `experiment` key under the `fluxsite` key and sets
the experiment key to be optional.

Fixes #192